Added rudimentary account creation and login functionality.
Added functions to the control class for making an account and logging in. createAccount() and login(). both take 2 String parameters, the username and the password, in that order. If trying to make an account and that username is already in use, returns false. If trying to login and no account with that username or the password if wrong, returns false. otherwise returns true for both.