Added to GUI + more
Added  "no current classes" and "no recent activity" labels to homepage, made the "view all classes" button on account page functional, and added a check to teacher class and control class so that if there are no classes in the teacher's class list, it will display "no classes". Otherwise, it will display the classes assigned to the user.